finish inital pst support PST file support: Impliments a scanner which extracts data from Outlook pst files The files are treated as a virtual file system. The scanner adds new entries to the filesystem database. The pst file can then browsed as though it was part of the filesystem. The filesystem part is implimented in FileSystem.py as a 'Virtual Filesystem' simillar to the Zip implimentation.
This allows email text, attachments, contact information etc to be scanned by the other
scanners such as the text indexer, virus scanner etc.

Furter processing in the scanner will probably be done in the future to build email/contact
specific tables, and then reports added to effectively browse and search them. For now, at
least we have access to the contents in a uncompressed/unencrypted form.