add a new method of type translation into the database for performance There is a new table option "data_callbacks" which takes a python function which translates between database types and display types. Unlike "callbacks" it allows searching (only [=><] etc). We also now support >= and <= searches.
This increases performance when filtering dates and ip addresses as the index
can be used rather than a full table scan.

See documentation in HTMLUI.py table function or details. This may have to be
ported to other UI's (?) and it also should fail more gracefully when a 'like'
search is performed.