convert a bunch of DATETIME columns to TIMESTAMP
In the same vein as disk forensics, this patch changes a bunch of fields
from DATETIME to TIMESTAMP. This has the following implications:

1. dates are now stored internally in mysql as UTC
2. everything should still work the same way since what goes in is what
comes out *unless* the timezone is changed in mysql.
3. In the future we may wish to display times in another timezone, this change
allows us to do this simply by changing the time_zone variable in mysql before
selecting the results.