Feature 3
You can no longer make a case that is the name of an existing database. This might cause some issues for tests etc? but tests should be dropping cases first anyway. This is important since what if I have a flagDB called foo and then try and make a case called foo. Bad things.